refactor(core): Use a rest controller for status endpoints
The Endpoint classes that are used to implement the status endpoint
have significantly changed in Spring Boot 2. Instead of creating and
exposing an endpoint, just use a rest controller.
