+++ rosco-web/src/main/groovy/com/netflix/spinnaker/rosco/controllers/StatusController.groovy
@@ -1,11 +1,11 @@
 /*
  * Copyright 2016 Schibsted ASA.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
+ * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,15 +14,19 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.rosco.endpoints
+package com.netflix.spinnaker.rosco.controllers
 
 import com.netflix.spinnaker.rosco.api.BakeStatus
 import com.netflix.spinnaker.rosco.persistence.BakeStore
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.stereotype.Component
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
 
-@Component
-class StatusHandler {
+@ConditionalOnProperty(value = "endpoints.status.enabled", matchIfMissing = true)
+@RestController
+@RequestMapping("/status")
+class StatusController {
 
   private final BakeStore bakeStore
   private final String roscoInstanceId
@@ -32,16 +36,18 @@ class StatusHandler {
   }
 
   @Autowired
-  StatusHandler(BakeStore bakeStore, String roscoInstanceId) {
+  StatusController(BakeStore bakeStore, String roscoInstanceId) {
     this.bakeStore = bakeStore
     this.roscoInstanceId = roscoInstanceId
   }
 
+  @RequestMapping("/instance")
   public Map<String, Object> instanceIncompleteBakes() {
     def instanceIncompleteBakeIds = bakeStore.getThisInstanceIncompleteBakeIds()
     return getBakesAndInstanceStatus(instanceIncompleteBakeIds)
   }
 
+  @RequestMapping("/all")
   public Map<String, Object> allIncompleteBakes() {
     def instances = [:]
     def allIncompleteBakeIds = bakeStore.getAllIncompleteBakeIds()
