--- rosco-web/src/main/groovy/com/netflix/spinnaker/rosco/endpoints/AllStatusEndpoint.groovy
@@ -1,42 +1,0 @@
-/*
- * Copyright 2016 Schibsted ASA.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.netflix.spinnaker.rosco.endpoints
-
-import groovy.util.logging.Slf4j
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.actuate.endpoint.AbstractEndpoint
-import org.springframework.boot.context.properties.ConfigurationProperties
-import org.springframework.stereotype.Component
-
-@Component
-@Slf4j
-@ConfigurationProperties(prefix = "endpoints.status", ignoreUnknownFields = false)
-class AllStatusEndpoint extends AbstractEndpoint<Map<String, Object>> {
-
-  private final StatusHandler statusHandler
-
-  @Autowired
-  AllStatusEndpoint(StatusHandler statusHandler) {
-    super("status", false, true)
-    this.statusHandler = statusHandler
-  }
-
-  @Override
-  public Map<String, Object> invoke() {
-    return statusHandler.allIncompleteBakes()
-  }
-}
