+++ rosco-web/src/test/groovy/com/netflix/spinnaker/rosco/controllers/AllStatusEndpointSpec.groovy
@@ -1,11 +1,11 @@
 /*
  * Copyright 2016 Schibsted ASA.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
+ * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.rosco.endpoints
+package com.netflix.spinnaker.rosco.controllers
 
 import com.google.common.collect.Sets
 import com.netflix.spinnaker.rosco.api.BakeStatus
+import com.netflix.spinnaker.rosco.controllers.StatusController
 import com.netflix.spinnaker.rosco.persistence.RedisBackedBakeStore
 import spock.lang.Specification
 import spock.lang.Subject
@@ -36,7 +37,7 @@ class AllStatusEndpointSpec extends Specification {
     def bakeStoreMock = Mock(RedisBackedBakeStore)
 
     @Subject
-    def statusHandler = new StatusHandler(bakeStoreMock, localInstanceId)
+    def statusHandler = new StatusController(bakeStoreMock, localInstanceId)
 
     when:
     def instances = statusHandler.allIncompleteBakes()
@@ -51,7 +52,7 @@ class AllStatusEndpointSpec extends Specification {
     setup:
     def bakeStoreMock = Mock(RedisBackedBakeStore)
     bakeStoreMock.getAllIncompleteBakeIds() >> new HashMap<String, Set<String>>()
-    def statusHandler = new StatusHandler(bakeStoreMock, localInstanceId)
+    def statusHandler = new StatusController(bakeStoreMock, localInstanceId)
 
     when:
     def instances = statusHandler.allIncompleteBakes()
@@ -64,7 +65,7 @@ class AllStatusEndpointSpec extends Specification {
   void 'no instances incomplete bakes'() {
     setup:
     def bakeStoreMock = Mock(RedisBackedBakeStore)
-    def statusHandler = new StatusHandler(bakeStoreMock, localInstanceId)
+    def statusHandler = new StatusController(bakeStoreMock, localInstanceId)
 
     when:
     def instances = statusHandler.allIncompleteBakes()
@@ -77,7 +78,7 @@ class AllStatusEndpointSpec extends Specification {
     setup:
     def bakeStoreMock = Mock(RedisBackedBakeStore)
     bakeStoreMock.getAllIncompleteBakeIds() >> { throw new RuntimeException() }
-    def statusHandler = new StatusHandler(bakeStoreMock, localInstanceId)
+    def statusHandler = new StatusController(bakeStoreMock, localInstanceId)
 
     when:
     statusHandler.allIncompleteBakes()
@@ -92,7 +93,7 @@ class AllStatusEndpointSpec extends Specification {
     bakeStoreMock.getAllIncompleteBakeIds() >> [(localInstanceId): Sets.newHashSet(LOCAL_JOB_ID), (remoteInstanceId): Sets.newHashSet(REMOTE_JOB_ID)]
     bakeStoreMock.retrieveBakeStatusById(LOCAL_JOB_ID) >> { throw new RuntimeException() }
     bakeStoreMock.retrieveBakeStatusById(REMOTE_JOB_ID) >> { throw new RuntimeException() }
-    def statusHandler = new StatusHandler(bakeStoreMock, localInstanceId)
+    def statusHandler = new StatusController(bakeStoreMock, localInstanceId)
 
     when:
     statusHandler.allIncompleteBakes()
