+++ rosco-web/src/test/groovy/com/netflix/spinnaker/rosco/controllers/InstanceStatusEndpointSpec.groovy
@@ -1,11 +1,11 @@
 /*
  * Copyright 2016 Schibsted ASA.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
+ * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.rosco.endpoints
+package com.netflix.spinnaker.rosco.controllers
 
 import com.google.common.collect.Sets
 import com.netflix.spinnaker.rosco.api.BakeStatus
+import com.netflix.spinnaker.rosco.controllers.StatusController
 import com.netflix.spinnaker.rosco.persistence.RedisBackedBakeStore
 import spock.lang.Specification
 import spock.lang.Subject
@@ -34,7 +35,7 @@ class InstanceStatusEndpointSpec extends Specification {
     def bakeStoreMock = Mock(RedisBackedBakeStore)
 
     @Subject
-    def statusHandler = new StatusHandler(bakeStoreMock, instanceId)
+    def statusHandler = new StatusController(bakeStoreMock, instanceId)
 
     when:
     def instanceInfo = statusHandler.instanceIncompleteBakes()
@@ -52,7 +53,7 @@ class InstanceStatusEndpointSpec extends Specification {
     bakeStoreMock.getThisInstanceIncompleteBakeIds() >> new HashSet<>()
 
     @Subject
-    def statusHandler = new StatusHandler(bakeStoreMock, instanceId)
+    def statusHandler = new StatusController(bakeStoreMock, instanceId)
 
     when:
     def instanceInfo = statusHandler.instanceIncompleteBakes()
@@ -68,7 +69,7 @@ class InstanceStatusEndpointSpec extends Specification {
     def bakeStoreMock = Mock(RedisBackedBakeStore)
 
     @Subject
-    def statusHandler = new StatusHandler(bakeStoreMock, instanceId)
+    def statusHandler = new StatusController(bakeStoreMock, instanceId)
 
     when:
     def instanceInfo = statusHandler.instanceIncompleteBakes()
@@ -84,7 +85,7 @@ class InstanceStatusEndpointSpec extends Specification {
     bakeStoreMock.getThisInstanceIncompleteBakeIds() >> { throw new RuntimeException() }
 
     @Subject
-    def statusHandler = new StatusHandler(bakeStoreMock, instanceId)
+    def statusHandler = new StatusController(bakeStoreMock, instanceId)
 
     when:
     statusHandler.instanceIncompleteBakes()
@@ -100,7 +101,7 @@ class InstanceStatusEndpointSpec extends Specification {
     bakeStoreMock.retrieveBakeStatusById(JOB_ID) >> { throw new RuntimeException() }
 
     @Subject
-    def statusHandler = new StatusHandler(bakeStoreMock, instanceId)
+    def statusHandler = new StatusController(bakeStoreMock, instanceId)
 
     when:
     statusHandler.instanceIncompleteBakes()
